refactor: feature/category 수정 및 적용
