+++ frontend/src/features/category/api/category.api.ts
@@ -1,4 +1,4 @@
-import { httpClient } from './http';
+import { httpClient } from '@/shared/utils/http';
 
 export const fetchCategory = async () => {
   const { data } = await httpClient.get('/category');
